get the index.js from home
